spider/vnf_examples/source/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-Classful==0.14.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
Pillow==8.4.0
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
kubernetes==19.15.0
